+++ iac/env/tts/dev/insert-state-info.sql
@@ -22,22 +22,22 @@ For dummy phone numbers use 555-0100 through 555-0199 which are reserved for fic
 BEGIN;
 
 INSERT INTO state_info(id, state, state_abbreviation, email, phone, region, email_cc)
-select * from (select '15' as id, 'Iowa' as state, 'IA' as state_abbreviation, 'IA-test@usda.example' as email, '1235550101' as phone, 'MWRO' AS region, 'IA-test-cc@usda.example' as email_cc) as temp
+select * from (select '15' as id, 'Iowa' as state, 'IA' as state_abbreviation, 'IA-test@agency.example' as email, '1235550101' as phone, 'MWRO' AS region, 'IA-test-cc@agency.example' as email_cc) as temp
 WHERE NOT EXISTS
 (select id from state_info where id = '15') limit 1;
 
 INSERT INTO state_info(id, state, state_abbreviation, email, phone, region, email_cc)
-select * from (select '18' as id, 'Louisiana' as state, 'LA' as state_abbreviation, 'LA-test@usda.example' as email, '1235550102' as phone, 'SWRO' AS region, 'LA-test-cc@usda.example' as email_cc) as temp
+select * from (select '18' as id, 'Louisiana' as state, 'LA' as state_abbreviation, 'LA-test@agency.example' as email, '1235550102' as phone, 'SWRO' AS region, 'LA-test-cc@agency.example' as email_cc) as temp
 WHERE NOT EXISTS
 (select id from state_info where id = '18') limit 1;
 
 INSERT INTO state_info(id, state, state_abbreviation, email, phone, region, email_cc)
-select * from (select '21' as id, 'Massachusetts' as state, 'MA' as state_abbreviation, 'MA-test@usda.example' as email, '1235550103' as phone, 'NERO' AS region, 'MA-test-cc@usda.example' as email_cc) as temp
+select * from (select '21' as id, 'Massachusetts' as state, 'MA' as state_abbreviation, 'MA-test@agency.example' as email, '1235550103' as phone, 'NERO' AS region, 'MA-test-cc@agency.example' as email_cc) as temp
 WHERE NOT EXISTS
 (select id from state_info where id = '21') limit 1;
 
 INSERT INTO state_info(id, state, state_abbreviation, email, phone, region, email_cc)
-select * from (select '26' as id, 'Montana' as state, 'MT' as state_abbreviation, 'MT-test@usda.example' as email, '1235550104' as phone, 'MPRO' AS region, 'MT-test-cc@usda.example' as email_cc) as temp
+select * from (select '26' as id, 'Montana' as state, 'MT' as state_abbreviation, 'MT-test@agency.example' as email, '1235550104' as phone, 'MPRO' AS region, 'MT-test-cc@agency.example' as email_cc) as temp
 WHERE NOT EXISTS
 (select id from state_info where id = '26') limit 1;
 
@@ -47,12 +47,12 @@ WHERE NOT EXISTS
 (select id from state_info where id = '101') limit 1;
 
 INSERT INTO state_info(id, state, state_abbreviation, email, phone, region, email_cc)
-select * from (select '102' as id, 'Echo Bravo' as state, 'EB' as state_abbreviation, 'EB-test@usda.example' as email, '1235550106' as phone, 'EB-MPRO' AS region, 'EB-test-cc@usda.example' as email_cc) as temp
+select * from (select '102' as id, 'Echo Bravo' as state, 'EB' as state_abbreviation, 'EB-test@agency.example' as email, '1235550106' as phone, 'EB-MPRO' AS region, 'EB-test-cc@agency.example' as email_cc) as temp
 WHERE NOT EXISTS
 (select id from state_info where id = '102') limit 1;
 
 INSERT INTO state_info(id, state, state_abbreviation, email, phone, region, email_cc)
-select * from (select '103' as id, 'Echo Charlie' as state, 'EC' as state_abbreviation, 'EC-test@usda.example' as email, '1235550107' as phone, 'EC-MPRO' AS region, 'EC-test-cc@usda.example' as email_cc) as temp
+select * from (select '103' as id, 'Echo Charlie' as state, 'EC' as state_abbreviation, 'EC-test@agency.example' as email, '1235550107' as phone, 'EC-MPRO' AS region, 'EC-test-cc@agency.example' as email_cc) as temp
 WHERE NOT EXISTS
 (select id from state_info where id = '103') limit 1;